# Sistema_vuelos.v1/vuelos-backend/openapi.yaml
openapi: 3.0.3
info:
  title: Vuelos API
  description: |
    API REST para el módulo de vuelos de la plataforma Booking académica.

    **Versionamiento**: todas las rutas usan el prefijo `/api/v1/`.

    **Autenticación**: Bearer JWT en el header `Authorization`.
    Obtener el token con `POST /api/v1/auth/login`.

    **Validación**: todos los endpoints mutantes (POST/PUT/PATCH) validan el cuerpo con
    Zod antes de ejecutar lógica de negocio. Errores de validación retornan HTTP 400.

    **Integridad de asientos**: la creación de reservas y la cancelación se ejecutan en
    transacciones Prisma atómicas — se garantiza que `availableSeats` nunca queda
    inconsistente aunque haya concurrencia.

    **Integracion con Booking Central**: el sistema expone contratos gRPC en los archivos
    `.proto` de `src/grpc/proto/`. OpenAPI documenta solamente el contrato REST/JSON;
    los metodos gRPC no se declaran aqui para no mezclar protocolos.
  version: 1.1.0
  contact:
    name: Equipo Vuelos
servers:
  - url: https://integracion-sistemas2026.onrender.com/api/v1
    description: Producción (Render)
  - url: http://localhost:3000/api/v1
    description: Desarrollo local

# ============================================================
#   SEGURIDAD GLOBAL
# ============================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido en POST /auth/login
  schemas:
...
    SuccessResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, example: true }
        data:    { type: object }
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          required: [code, message]
          properties:
            code:    { type: string, example: VALIDATION }
            message: { type: string, example: Descripción del error }

    # ── Entidades ────────────────────────────────────────────────
...
    Country:
      type: object
      properties:
        id:           { type: string }
        name:         { type: string, example: Ecuador }
        isoCode:      { type: string, example: EC, minLength: 2, maxLength: 3 }
        phoneCode:    { type: string, nullable: true }
        currencyCode: { type: string, nullable: true }
    City:
      type: object
      properties:
        id:       { type: string }
        name:     { type: string, example: Quito }
        iataCode: { type: string, nullable: true, example: UIO }
        country:  { $ref: '#/components/schemas/Country' }
    Airport:
      type: object
      properties:
        id:       { type: string }
        iataCode: { type: string, example: UIO }
        name:     { type: string, example: Aeropuerto Mariscal Sucre }
        timezone: { type: string, example: America/Guayaquil }
        city:     { $ref: '#/components/schemas/City' }
    Airline:
      type: object
      properties:
        id:        { type: string }
        iataCode:  { type: string, example: AV }
        name:      { type: string, example: Avianca }
        logoUrl:   { type: string, nullable: true }
        country:   { $ref: '#/components/schemas/Country' }
...
    Segment:
      type: object
      properties:
        id:                  { type: string }
        segmentNumber:       { type: string, example: AV101 }
        flightId:            { type: string, nullable: true }
        originAirportId:     { type: string }
        destinationAirportId: { type: string }
        departureDateTime:   { type: string, format: date-time, example: '2025-06-15T08:30:00Z' }
        arrivalDateTime:     { type: string, format: date-time, example: '2025-06-15T10:15:00Z' }
        airlineId:           { type: string }
        aircraftId:          { type: string, nullable: true }
        estimatedDuration:   { type: integer, description: Duración en minutos, example: 105 }
        originAirport:       { $ref: '#/components/schemas/Airport' }
        destinationAirport:  { $ref: '#/components/schemas/Airport' }
        airline:             { $ref: '#/components/schemas/Airline' }
    FlightClass:
      type: object
      properties:
        id:             { type: string }
        flightId:       { type: string }
        cabinClass:     { type: string, enum: [ECONOMY, PREMIUM_ECONOMY, BUSINESS, FIRST] }
        availableSeats: { type: integer, minimum: 0 }
        basePrice:      { type: number, format: decimal, example: 89.00 }
        flight:
          nullable: true
          $ref: '#/components/schemas/Flight'
    Flight:
      type: object
      properties:
        id:                     { type: string }
        originAirportIata:      { type: string, example: UIO }
        destinationAirportIata: { type: string, example: BOG }
        departureDate:          { type: string, format: date }
        status:                 { type: string, enum: [SCHEDULED, DELAYED, CANCELLED, COMPLETED] }
        # Campos calculados por FlightMapper a partir de segmentos
        flightNumber:      { type: string, nullable: true }
        departureDateTime: { type: string, format: date-time, nullable: true }
        arrivalDateTime:   { type: string, format: date-time, nullable: true }
        duration:          { type: integer, description: Duración total en minutos, nullable: true }
        stops:             { type: integer, nullable: true }
        lowestPrice:       { type: number, nullable: true }
        airline:           { nullable: true, $ref: '#/components/schemas/Airline' }
        segments:
          type: array
          items: { $ref: '#/components/schemas/Segment' }
        flightClasses:
          type: array
          items: { $ref: '#/components/schemas/FlightClass' }
    Passenger:
      type: object
      required: [firstName, lastName, documentNumber]
      properties:
        id:             { type: string }
        firstName:      { type: string, example: Juan }
        lastName:       { type: string, example: Pérez }
        documentNumber: { type: string, example: '1234567890' }
        seatNumber:     { type: string, nullable: true, example: 14C }
        flightClassId:  { type: string }
    Promotion:
      type: object
      properties:
        id:            { type: string }
        code:          { type: string, example: VERANO25 }
        discountType:  { type: string, enum: [PERCENTAGE, FIXED_AMOUNT] }
        discountValue: { type: number, example: 15 }
        maxUsages:     { type: integer, nullable: true }
        currentUsages: { type: integer }
        isActive:      { type: boolean }
    Reservation:
      type: object
      properties:
        id:              { type: string }
        reservationCode: { type: string, example: A3F7B2C1 }
        status:          { type: string, enum: [PENDING, CONFIRMED, CANCELLED, COMPLETED] }
        totalAmount:     { type: number, format: decimal }
        createdAt:       { type: string, format: date-time }
        userId:          { type: string }
        flightId:        { type: string }
        promotionId:     { type: string, nullable: true }
        flight:
          nullable: true
          $ref: '#/components/schemas/Flight'
        passengers:
          type: array
          items: { $ref: '#/components/schemas/Passenger' }
        promotion:
          nullable: true
          $ref: '#/components/schemas/Promotion'
        user:
          nullable: true
          type: object
          properties:
            id:            { type: string }
            email:         { type: string }
            firstName:     { type: string }
            firstLastName: { type: string }
...
  responses:
    Unauthorized:
      description: Token ausente o inválido
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Sin permisos para esta operación
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    BadRequest:
      description: Datos de entrada inválidos (fallo de validación Zod)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Conflicto de datos (duplicado o restricción de negocio)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }

# ============================================================
#   PATHS
# ============================================================
paths:
...
  /reservations:
    post:
      tags: [Reservations]
      summary: Crear reserva
      description: |
        Transacción atómica que:
        1. Verifica disponibilidad (`availableSeats >= pasajeros`)
        2. Aplica descuento si se envía `promotionCode` (valida existencia, estado activo y límite de usos)
        3. Crea la reserva y sus pasajeros en un solo commit
        4. Decrementa `availableSeats` e incrementa `currentUsages` de la promo

        Si dos peticiones concurrentes compiten por el último asiento, solo una confirma.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [flightClassId, passengers]
              properties:
                flightClassId:
                  type: string
                  format: uuid
                passengers:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [firstName, lastName, documentNumber]
                    properties:
                      firstName:      { type: string }
                      lastName:       { type: string }
                      documentNumber: { type: string }
                      seatNumber:     { type: string, nullable: true }
                promotionCode:
                  type: string
                  example: VERANO25
      responses:
        '201':
          description: Reserva confirmada
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Reservation' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422':
          description: Sin disponibilidad
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    get:
      tags: [Reservations]
      summary: Listar todas las reservas (admin)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Lista de reservas
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Reservation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /reservations/my:
    get:
      tags: [Reservations]
      summary: Mis reservas
      description: Devuelve el historial completo del usuario autenticado, con vuelo, pasajeros y promo.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Reservas del usuario
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Reservation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /reservations/flight-classes/{flightClassId}/occupied-seats:
    parameters:
      - { name: flightClassId, in: path, required: true, schema: { type: string }, description: ID de la clase de vuelo }
    get:
      tags: [Reservations]
      summary: Asientos ocupados por clase de vuelo
      description: Devuelve codigos de asiento asignados para pintar el mapa y bloquear seleccion.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Lista de asientos ocupados
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { type: string, example: 14C }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /reservations/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [Reservations]
      summary: Detalle de una reserva
      description: Solo el dueño o un ADMIN pueden acceder.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Reserva encontrada
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Reservation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Reservations]
      summary: Cancelar reserva (DELETE)
      description: |
        Cancela la reserva y en una sola transacción atómica:
        - Marca el estado como `CANCELLED`
        - Restaura los asientos en `FlightClass.availableSeats`
        - Decrementa `Promotion.currentUsages` si la reserva usó un código
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Reserva cancelada
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          cancelled:       { type: boolean }
                          reservationCode: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /reservations/{id}/cancel:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    patch:
      tags: [Reservations]
      summary: Cancelar reserva (PATCH)
      description: |
        Equivalente funcional a `DELETE /{id}` — existe para compatibilidad con
        el cliente Angular que usa PATCH. Aplica la misma lógica atómica de
        restauración de asientos y decremento de promo.
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: Reserva cancelada
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          cancelled:       { type: boolean }
                          reservationCode: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /reservations/{id}/passengers/{passengerId}/seat:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string }, description: ID de la reserva }
      - { name: passengerId, in: path, required: true, schema: { type: string }, description: ID del pasajero }
    patch:
      tags: [Reservations]
      summary: Asignar asiento a un pasajero
      description: |
        Permite asignar o cambiar el número de asiento de un pasajero.
        - Solo el dueño de la reserva o un ADMIN pueden llamar este endpoint.
        - Retorna 409 si el asiento ya está ocupado por otro pasajero del mismo vuelo y clase.
        - El número de asiento se normaliza a mayúsculas (ej: `14c` → `14C`).

        Se puede llamar durante el check-in o de forma independiente post check-in.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [seatNumber]
              properties:
                seatNumber:
                  type: string
                  example: 14C
                  description: Número de asiento (letras y números, se convierte a mayúsculas)
      responses:
        '200':
          description: Asiento asignado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Passenger' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }

  # ── PAYMENTS ────────────────────────────────────────────────
...
  /admin/reservations:
    get:
      tags: [Admin]
      summary: Listar reservas (admin CRUD)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Lista, content: { application/json: { schema: { $ref: '#/components/schemas/SuccessResponse' } } } }
...